# src/main/resources/openapi.json
{"openapi":"3.1.0","info":{"title":"SAO Bot Database API","description":"Документация для SAO Bot Database API","contact":{"name":"Knalis","url":"https://github.com/knliiss","email":"vitallot21@gmail.com"},"license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"v1.0.0"},"servers":[{"url":"http://localhost:8090","description":"Generated server url"}],"tags":[{"name":"User Configurations","description":"API для управления конфигурациями настроек пользователя"},{"name":"User Message Packs","description":"API для работы с покупкой и активацией message packs у пользователя"},{"name":"Gangs","description":"API для работы с бандами"},{"name":"Users","description":"API для работы с пользователями"},{"name":"Scheduled Messages","description":"API для работы с запланированными сообщениями"},{"name":"Drops","description":"API для работы с дропами игроков"},{"name":"Message Packs","description":"API для работы с пакетами сообщений"},{"name":"User Settings","description":"API для работы с настройками уведомлений пользователя"}],"paths":{"/api/v1/users/{userId}/settings":{"get":{"tags":["User Settings"],"summary":"Получить настройки пользователя","description":"Возвращает текущие настройки уведомлений для указанного пользователя","operationId":"getSettings","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SettingsDTO"}}}}}},"put":{"tags":["User Settings"],"summary":"Обновить все настройки пользователя","description":"Полностью обновляет все настройки уведомлений для указанного пользователя","operationId":"updateAllSettings","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/users/{id}":{"get":{"tags":["Users"],"summary":"Получить пользователя по ID","description":"Возвращает DTO пользователя по его ID","operationId":"getUser","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"put":{"tags":["Users"],"summary":"Обновить пользователя","description":"Обновляет данные пользователя по ID","operationId":"updateUser","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"nickname","in":"query","description":"Новый никнейм (опционально)","required":false,"schema":{"type":"string"}},{"name":"messagePackId","in":"query","description":"ID нового message pack (опционально)","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"delete":{"tags":["Users"],"summary":"Удалить пользователя","description":"Удаляет пользователя по ID","operationId":"deleteUser","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/scheduled-messages/{id}":{"put":{"tags":["Scheduled Messages"],"summary":"Обновить запланированное сообщение по ID","description":"Обновляет запланированное сообщение с указанным ID","operationId":"updateScheduledMessage","parameters":[{"name":"id","in":"path","description":"ID запланированного сообщения","required":true,"schema":{"type":"string"}},{"name":"content","in":"query","description":"Обновленный текст(опционально","required":false,"schema":{"type":"string"}},{"name":"scheduledTime","in":"query","description":"Обновленное время отправки(опционально","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScheduledMessage"}}}}}},"delete":{"tags":["Scheduled Messages"],"summary":"Удалить запланированное сообщение по ID","description":"Удаляет запланированное сообщение с указанным ID","operationId":"deleteScheduledMessage","parameters":[{"name":"id","in":"path","description":"ID запланированного сообщения","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/users":{"get":{"tags":["Users"],"summary":"Получить список пользователей","description":"Возвращает список всех пользователей или фильтрует по настройке уведомлений","operationId":"getUsers","parameters":[{"name":"setting","in":"query","description":"Название настройки уведомления для фильтрации","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"post":{"tags":["Users"],"summary":"Создать пользователя","description":"Создает нового пользователя и возвращает его DTO","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/v1/users/{userId}/config":{"get":{"tags":["User Configurations"],"summary":"Получить конфигурации пользователя","description":"Возвращает список всех конфигураций настроек пользователя. Если параметр `active=true`, возвращается только активная конфигурация.","operationId":"getUserConfigs","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"active","in":"query","description":"Если true — вернуть только активную конфигурацию","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConfigDTO"}}}}}}},"post":{"tags":["User Configurations"],"summary":"Создать новую конфигурацию настроек","description":"Создаёт новую конфигурацию для пользователя с заданным именем.","operationId":"createUserConfig","parameters":[{"name":"userId","in":"path","description":"ID пользователя, для которого создаётся конфигурация","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","description":"Имя новой конфигурации"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ConfigDTO"}}}}}}},"/api/v1/users/{id}/balance/add/{amount}":{"post":{"tags":["Users"],"summary":"Добавить баланс пользователю","description":"Добавляет указанное количество монет пользователю","operationId":"addBalance","parameters":[{"name":"id","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"amount","in":"path","description":"Сумма для добавления","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/v1/users/balance/add-all/{amount}":{"post":{"tags":["Users"],"summary":"Добавить баланс всем пользователям","description":"Добавляет указанную сумму всем пользователям","operationId":"addBalanceToAll","parameters":[{"name":"amount","in":"path","description":"Сумма для добавления","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/user/message-pack/buy":{"post":{"tags":["User Message Packs"],"summary":"Купить message pack","description":"Позволяет пользователю купить и активировать message pack","operationId":"buyPack","parameters":[{"name":"userId","in":"query","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"packId","in":"query","description":"ID message pack","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/scheduled-messages":{"get":{"tags":["Scheduled Messages"],"summary":"Получить запланированные сообщения","description":"Возвращает список всех запланированных сообщений, можно фильтровать по Id пользователя","operationId":"getScheduledMessages","parameters":[{"name":"userId","in":"query","description":"Id пользователя(Опционально)","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduledMessage"}}}}}}},"post":{"tags":["Scheduled Messages"],"summary":"Создать запланированное сообщение","description":"Создаёт новое запланированное сообщение","operationId":"createScheduledMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduledMessage"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScheduledMessage"}}}}}}},"/api/v1/gangs":{"get":{"tags":["Gangs"],"summary":"Получить все банды","description":"Возвращает список всех банд","operationId":"getAllGangs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GangDTO"}}}}}}},"post":{"tags":["Gangs"],"summary":"Создать банду","description":"Создаёт новую банду от имени пользователя","operationId":"createGang","parameters":[{"name":"userId","in":"query","description":"ID пользователя, создающего банду","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"name","in":"query","description":"Название банды","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GangDTO"}}}}}}},"/api/v1/gangs/{gangId}/transfer":{"post":{"tags":["Gangs"],"summary":"Передать владение бандой","description":"Передаёт владение бандой другому пользователю","operationId":"transferOwnership","parameters":[{"name":"gangId","in":"path","description":"ID банды","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"newOwnerId","in":"query","description":"ID нового владельца","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/gangs/{gangId}/members":{"post":{"tags":["Gangs"],"summary":"Добавить участника в банду","description":"Добавляет пользователя в банду","operationId":"addMember","parameters":[{"name":"gangId","in":"path","description":"ID банды","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"actorId","in":"query","description":"ID пользователя, выполняющего действие","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"targetId","in":"query","description":"ID пользователя, которого добавляют","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/gangs/{gangId}/join":{"post":{"tags":["Gangs"],"summary":"Вступить в банду","description":"Позволяет пользователю вступить в указанную банду","operationId":"joinGang","parameters":[{"name":"gangId","in":"path","description":"ID банды","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"query","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/gangs/leave":{"post":{"tags":["Gangs"],"summary":"Покинуть банду","description":"Позволяет пользователю покинуть свою банду","operationId":"leaveGang","parameters":[{"name":"userId","in":"query","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/drops":{"get":{"tags":["Drops"],"summary":"Получить все дропы","description":"Возвращает список всех дропов. Можно фильтровать по имени игрока и названию предмета.","operationId":"getAllDrops","parameters":[{"name":"username","in":"query","description":"Имя игрока для фильтрации","required":false,"schema":{"type":"string"}},{"name":"itemName","in":"query","description":"Название предмета для фильтрации","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Drop"}}}}}}},"post":{"tags":["Drops"],"summary":"Добавить новый дроп","description":"Сохраняет новый дроп и возвращает обновлённый объект PlayerDrops.","operationId":"addDrop","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Drop"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayerDrops"}}}}}}},"/api/v1/users/{userId}/settings/{type}":{"patch":{"tags":["User Settings"],"summary":"Переключить настройку уведомления","description":"Переключает указанную настройку уведомлений пользователя","operationId":"toggleSetting","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"type","in":"path","description":"Тип настройки уведомления","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/users/{userId}/config/activate/{configId}":{"patch":{"tags":["User Configurations"],"summary":"Активировать конфигурацию","description":"Делает указанную конфигурацию активной для данного пользователя.","operationId":"setActiveConfig","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"configId","in":"path","description":"ID конфигурации, которую нужно активировать","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/users/by-nickname/{nickname}":{"get":{"tags":["Users"],"summary":"Получить пользователя по никнейму","description":"Возвращает DTO пользователя по его никнейму","operationId":"getUserByNickname","parameters":[{"name":"nickname","in":"path","description":"Никнейм пользователя","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/v1/message-packs":{"get":{"tags":["Message Packs"],"summary":"Получить все пакеты сообщений","description":"Возвращает список всех пакетов сообщений, можно фильтровать по редкости","operationId":"getAll","parameters":[{"name":"rarity","in":"query","description":"Фильтр по редкости","required":false,"schema":{"type":"string","enum":["COMMON","UNCOMMON","RARE","EPIC","LEGENDARY","CUSTOM"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessagePackDTO"}}}}}}}},"/api/v1/message-packs/{id}":{"get":{"tags":["Message Packs"],"summary":"Получить пакет сообщений по ID","description":"Возвращает пакет сообщений с указанным ID","operationId":"getById","parameters":[{"name":"id","in":"path","description":"ID пакета сообщений","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessagePackDTO"}}}}}}},"/api/v1/message-packs/{id}/messages/{key}":{"get":{"tags":["Message Packs"],"summary":"Получить сообщение из пакета по ключу","description":"Возвращает конкретное сообщение из пакета по ключу","operationId":"getMessage","parameters":[{"name":"id","in":"path","description":"ID пакета сообщений","required":true,"schema":{"type":"string"}},{"name":"key","in":"path","description":"Ключ сообщения","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/message-packs/batch":{"get":{"tags":["Message Packs"],"summary":"Получить несколько пакетов сообщений по списку ID","description":"Возвращает список пакетов сообщений по переданным ID","operationId":"getByIds","parameters":[{"name":"ids","in":"query","description":"Список ID пакетов сообщений","required":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessagePackDTO"}}}}}}}},"/api/v1/gangs/{gangId}":{"get":{"tags":["Gangs"],"summary":"Получить банду по ID","description":"Возвращает информацию о банде с указанным ID","operationId":"getGang","parameters":[{"name":"gangId","in":"path","description":"ID банды","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GangDTO"}}}}}}},"/api/v1/users/{userId}/config/{configId}":{"delete":{"tags":["User Configurations"],"summary":"Удалить конфигурацию","description":"Удаляет указанную конфигурацию настроек пользователя.","operationId":"deleteUserConfig","parameters":[{"name":"userId","in":"path","description":"ID пользователя","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"configId","in":"path","description":"ID конфигурации для удаления","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/gangs/{gangId}/members/{userId}":{"delete":{"tags":["Gangs"],"summary":"Удалить участника из банды","description":"Удаляет указанного пользователя из банды","operationId":"removeMember","parameters":[{"name":"gangId","in":"path","description":"ID банды","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"actorId","in":"query","description":"ID пользователя, выполняющего действие","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userId","in":"path","description":"ID пользователя, которого удаляют","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"SettingsDTO":{"type":"object","description":"DTO с новыми настройками","properties":{"messagePackId":{"type":"string"},"notificationSettings":{"type":"object","additionalProperties":{"type":"boolean"}}}},"SubscriptionDTO":{"type":"object","properties":{"plan":{"type":"string"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"}}},"UserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"nickname":{"type":"string"},"balance":{"type":"number","format":"double"},"configId":{"type":"integer","format":"int64"},"subscription":{"$ref":"#/components/schemas/SubscriptionDTO"},"roles":{"type":"array","items":{"type":"string"}},"additionalAccounts":{"type":"array","items":{"type":"string"}},"gangId":{"type":"integer","format":"int64"}}},"ScheduledMessage":{"type":"object","description":"Данные запланированного сообщения","properties":{"id":{"type":"string"},"scheduledTime":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"message":{"type":"string"}}},"UserCreateRequest":{"type":"object","description":"DTO с данными для создания пользователя","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"nickname":{"type":"string"}}},"ConfigDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"settings":{"$ref":"#/components/schemas/SettingsDTO"}}},"GangDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"open":{"type":"boolean"},"memberLimit":{"type":"integer","format":"int32"},"owner":{"$ref":"#/components/schemas/GangMemberDTO"},"members":{"type":"array","items":{"$ref":"#/components/schemas/GangMemberDTO"}}}},"GangMemberDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"nickname":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"Drop":{"type":"object","description":"Данные дропа для добавления","properties":{"username":{"type":"string"},"item":{"type":"string"},"date":{"type":"string","format":"date-time"}}},"PlayerDrops":{"type":"object","properties":{"username":{"type":"string"},"drops":{"type":"array","items":{"$ref":"#/components/schemas/Drop"}}}},"MessagePackDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"emoji":{"type":"string"},"messages":{"type":"object","additionalProperties":{"type":"string"}},"cost":{"type":"number","format":"double"},"rarity":{"type":"string"}}}}}}